requirements.txt
fastapi[standard]
httpx
sqlalchemy
asyncpg
alembic
pydantic-settings
python-dotenv
pytest
pytest-asyncio
